changed data returned by codify_includes
codify_includes was implicitly joining the array of includes as a comma-deliniated
list and the commas were breaking the resulting script from being able to
run. I changed this to explicitly join the array on a new line ("\n"). I
also removed the unexplained ellipsis from line 153, because I've never
seen that syntax, and node apparently hasn't either.